Adding word splitting.

+++ src/Ironhide.Api.Host/IEncodingAlgorithm.cs
@@ -1,7 +1,9 @@
+using System;
+
 namespace Ironhide.Api.Host
 {
     public interface IEncodingAlgorithm
     {
-        string Encode(double startingNumber, string[] words);
+        string Encode(Int64 startingNumber, string[] words);
     }
 }
