+++ src/Ironhide.Api.Host/INumberSeriesGenerator.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace Ironhide.Api.Host
@@ -5,6 +6,6 @@ namespace Ironhide.Api.Host
     public interface INumberSeriesGenerator
     {
         IEnumerable<double> Generate(int count);
-        double GetNext(double previousNumber);
+        Int64 GetNext(double previousNumber);
     }
 }
